Have a basic template rendering in action
Following kind of what I know from using Express.js and handlebars.

A reusable main template resides in a layouts folder
and the remaining templates act as views and reside at the root of the
templates folder.

This allows us to not have to continually repeat the same code
to create our  HTML document, head and body tags. We will dynamically
insert the content of the view into the main template. We
can check the current view to conditionally render certain elements if
needed. For example in the login view we can hide the navigation bar