Created a basic login form for the landing page
Since this is an internal DBMS we don't need a real home page.
Refactored the loaders to use if statements for now because their
implementation is basic and will be changed later.

Added a hero image and created a CSS loader so we can dynamically
load CSS files. This should allow us to have a cleaner code base