Admin login functionality is there
Upon system start, a default adminUser is created from details in the
.env file. This user can be used to login to the admin panel. If the
admin table has records, then this process is skipped.

A user controller was created to keep all the logic for modifying
records in the DB in one place. This controller is used by the seeds,
and the login verification process at the moment but will be the main
controller for all Admin table related operations.

Now that we can authenticate users, it is time to create the admin
dashboard.