Basic layout for the dashboard has been set up. No content has been
added as we are testing the responsiveness of the layout.

A sidebar will be used for navigation but on mobile devices it will
become a horizontal nav on the bottom of the page in similar fashion to
mobile applications.

There is a topbar that displays the date and time and can serve
as a place for notifications if needed.