Refactored the code for getting all employees, roles, and departments.
Last night I was focused on rendering the data. Today I refactored the
repetitive code for the view templates. A helper folder was created,
this folder will contain helpers related to rendering content rather
than serve as a utility folder, which is already in place.

Next up is to handle the department, employee, and role options.