Can add departments
The logic is the same for adding roles so a reusable function has
been created. A handler has been implemented for adding departments,
roles, and employees. The role logic is implemented, just needs its own
view. The employee logic is not yet implemented and will be done at a
later date.